+++ playroom.config.js
@@ -29,7 +29,7 @@ const config = {
     snippets: './playroom/snippets.tsx',
     frameComponent: './playroom/frame-component.tsx',
     scope: './playroom/use-scope.tsx',
-    widths: [320, 768, 1024, 1366],
+    widths: [320, 360, 768, 1024, 1366],
     exampleCode,
     webpackConfig: () => ({
         module: {