+++ src/__stories__/loading-bar-story.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import {ThemeVariant, LoadingBar as LoadingBarComponent, Switch, useTheme, Box} from '..';
+import LoadingBarComponent from '../loading-bar';
+import {StorySection} from './helpers';
 
 export default {
     title: 'Components/Feedbacks/LoadingBar',
@@ -8,25 +9,8 @@ export default {
     },
 };
 
-const BackgroundTheme: React.FC = ({children}) => {
-    const [isInverseVariant, setIsInverseVariant] = React.useState(false);
-    const {colors} = useTheme();
-    return (
-        <ThemeVariant isInverse={isInverseVariant}>
-            <div style={{background: isInverseVariant ? colors.backgroundBrand : colors.background}}>
-                <Box paddingY={32} paddingX={16}>
-                    {children}
-                    <Switch name="isInverse" checked={isInverseVariant} onChange={setIsInverseVariant}>
-                        Inverse variant
-                    </Switch>
-                </Box>
-            </div>
-        </ThemeVariant>
-    );
-};
-
 export const LoadingBar: StoryComponent = () => (
-    <BackgroundTheme>
+    <StorySection title="Loading Bar">
         <LoadingBarComponent visible />
-    </BackgroundTheme>
+    </StorySection>
 );