+++ src/__stories__/use-element-dimensions-story.tsx
@@ -1,12 +1,28 @@
 import * as React from 'react';
 import {StorySection} from './helpers';
-import {useElementDimensions, Text2} from '..';
+import {useElementDimensions, Text2, ButtonPrimary} from '..';
 
 export default {
     title: 'Hooks/useElementDimensions',
 };
 
+const SomeComponent = ({innerRef}: {innerRef?: (node: HTMLElement | null) => void}) => {
+    return (
+        <div
+            ref={innerRef}
+            style={{
+                backgroundColor: 'gray',
+                width: 100,
+                height: 100,
+                resize: 'both',
+                overflow: 'auto',
+            }}
+        ></div>
+    );
+};
+
 export const UseElementSize: StoryComponent = () => {
+    const [isVisible, setIsVisible] = React.useState(true);
     const {height, width, ref} = useElementDimensions();
 
     return (
@@ -14,16 +30,12 @@ export const UseElementSize: StoryComponent = () => {
             <Text2 regular>
                 Height: {height}px, Width: {width}px
             </Text2>
-            <div
-                ref={ref}
-                style={{
-                    backgroundColor: 'gray',
-                    width: 100,
-                    height: 100,
-                    resize: 'both',
-                    overflow: 'auto',
-                }}
-            />
+            <br />
+            {isVisible && <SomeComponent innerRef={ref} />}
+            <br />
+            <ButtonPrimary onPress={() => setIsVisible((prev) => !prev)}>
+                {isVisible ? 'Unmount' : 'Mount'}
+            </ButtonPrimary>
         </StorySection>
     );
 };