+++ src/skins/types.tsx
@@ -58,7 +58,9 @@ export type Colors = {
     controlError: string;
     loadingBar: string;
     loadingBarBackground: string;
+    /** @deprecated */
     loadingBarBackgroundInverse: string;
+    /** @deprecated */
     loadingBarInverse: string;
 
     toggleAndroidInactive: string;