+++ src/skins/vivo.tsx
@@ -33,7 +33,7 @@ export const palette = {
     darkModeBlack: '#191919',
     darkModeGrey: '#242424',
     darkModeGrey6: '#313235',
-};
+} as const;
 
 export const getVivoSkin: GetSkin = () => {
     return {